+++ CHANGELOG.md
@@ -17,7 +17,8 @@ The types of changes are:
 
 ## [Unreleased](https://github.com/ethyca/fidesplus/compare/2.48.0...main)
 
-
+### Fixed
+- Fixed Snowflake DSR integration failing with syntax error [#5417](https://github.com/ethyca/fides/pull/5417)
 
 ## [2.48.0](https://github.com/ethyca/fidesplus/compare/2.47.1...2.48.0)